refactor(p17-m3): extract life store schema boundary
