+++ .github/workflows/architecture-gate.yml
@@ -67,5 +67,8 @@ jobs:
       - name: Run P17 M3-01 life store connection regression
         run: python tests/test_life_store_p17_m3_01.py -v
 
+      - name: Run P17 M3-02 life store schema regression
+        run: python tests/test_life_store_p17_m3_02.py -v
+
       - name: Compile P17 M2 and M3 seams
-        run: python -m py_compile app/backend/tiangong-backend/v3/zongdiaodu.py app/backend/tiangong-backend/v3/runtime_bootstrap.py app/backend/tiangong-backend/v3/runtime_composition.py app/backend/tiangong-backend/v3/runtime_lifecycle.py app/backend/tiangong-backend/v3/runtime_turn_orchestration.py app/backend/tiangong-backend/v3/runtime_tool_result_boundary.py src/life_service/embedded_runtime.py src/life_service/embedded_runtime_lifecycle.py src/life_service/embedded_runtime_wiring.py src/life_service/store.py src/life_service/store_connection.py src/total_gateway/runtime.py
+        run: python -m py_compile app/backend/tiangong-backend/v3/zongdiaodu.py app/backend/tiangong-backend/v3/runtime_bootstrap.py app/backend/tiangong-backend/v3/runtime_composition.py app/backend/tiangong-backend/v3/runtime_lifecycle.py app/backend/tiangong-backend/v3/runtime_turn_orchestration.py app/backend/tiangong-backend/v3/runtime_tool_result_boundary.py src/life_service/embedded_runtime.py src/life_service/embedded_runtime_lifecycle.py src/life_service/embedded_runtime_wiring.py src/life_service/store.py src/life_service/store_connection.py src/life_service/store_schema.py src/total_gateway/runtime.py
