+++ app/life-service/runtime314/life_service/.tiangong-generated-source.json
@@ -1,8 +1,8 @@
 {
-  "file_count": 54,
+  "file_count": 55,
   "mapping_id": "life-service-runtime",
   "schema": "tiangong.generated-source-marker.v1",
   "source": "src/life_service",
-  "tree_sha256": "6f7346960918cd3d0294fce2bf98be3f9e748bf083600f35c5af376f5ac35271",
+  "tree_sha256": "9e0dcfdf445d92bfde549ff866b46dd1d3c5cb93ebd625f571f457e660d9ac16",
   "warning": "Generated mirror. Edit the authoritative source and run scripts/sync-generated-sources.py --write."
 }
